client/requirements.txt
grpcio==1.58.0
grpcio-tools==1.58.0
protobuf==4.24.3
redis==5.0.0
